--- packages/coding-agent/src/slash-commands/acp-builtins/todo-state.ts
@@ -1,95 +1,0 @@
-import type { TodoPhase } from "../../tools/todo-write";
-import { getLatestTodoPhasesFromEntries, USER_TODO_EDIT_CUSTOM_TYPE } from "../../tools/todo-write";
-import type { AcpBuiltinCommandRuntime } from "./types";
-
-export interface TodoTaskMatch {
-	task: { content: string; status: string };
-	phase: TodoPhase;
-}
-
-export function tokenize(input: string): string[] {
-	const tokens: string[] = [];
-	let current = "";
-	let inQuote = false;
-	for (let index = 0; index < input.length; index++) {
-		const ch = input[index];
-		if (ch === "\\" && index + 1 < input.length) {
-			current += input[++index];
-			continue;
-		}
-		if (ch === '"') {
-			inQuote = !inQuote;
-			continue;
-		}
-		if (!inQuote && /\s/.test(ch)) {
-			if (current) {
-				tokens.push(current);
-				current = "";
-			}
-			continue;
-		}
-		current += ch;
-	}
-	if (current) tokens.push(current);
-	return tokens;
-}
-
-export function titleCase(text: string): string {
-	return text
-		.split(/\s+/)
-		.filter(Boolean)
-		.map(word => word[0].toUpperCase() + word.slice(1))
-		.join(" ");
-}
-
-export function titleCaseSentence(text: string): string {
-	const trimmed = text.trim();
-	if (!trimmed) return trimmed;
-	return trimmed[0].toUpperCase() + trimmed.slice(1);
-}
-
-export function findPhaseFuzzy(phases: TodoPhase[], query: string): TodoPhase | undefined {
-	const normalizedQuery = query.trim().toLowerCase();
-	if (!normalizedQuery) return undefined;
-	const exact = phases.find(phase => phase.name.toLowerCase() === normalizedQuery);
-	if (exact) return exact;
-	const prefixMatches = phases.filter(phase => phase.name.toLowerCase().startsWith(normalizedQuery));
-	if (prefixMatches.length === 1) return prefixMatches[0];
-	const substringMatches = phases.filter(phase => phase.name.toLowerCase().includes(normalizedQuery));
-	if (substringMatches.length === 1) return substringMatches[0];
-	return undefined;
-}
-
-export function findTaskFuzzy(phases: TodoPhase[], query: string): TodoTaskMatch | undefined {
-	const normalizedQuery = query.trim().toLowerCase();
-	if (!normalizedQuery) return undefined;
-	for (const phase of phases) {
-		for (const task of phase.tasks) {
-			if (task.content.toLowerCase() === normalizedQuery) return { task, phase };
-		}
-	}
-	const matches: TodoTaskMatch[] = [];
-	for (const phase of phases) {
-		for (const task of phase.tasks) {
-			if (task.content.toLowerCase().includes(normalizedQuery)) matches.push({ task, phase });
-		}
-	}
-	if (matches.length === 1) return matches[0];
-	const active = matches.filter(match => match.task.status === "in_progress" || match.task.status === "pending");
-	if (active.length === 1) return active[0];
-	return undefined;
-}
-
-export function currentPhases(runtime: AcpBuiltinCommandRuntime): TodoPhase[] {
-	const fromEntries = getLatestTodoPhasesFromEntries(runtime.sessionManager.getBranch());
-	return fromEntries.length > 0 ? fromEntries : runtime.session.getTodoPhases();
-}
-
-export function commitTodos(runtime: AcpBuiltinCommandRuntime, phases: TodoPhase[]): void {
-	runtime.session.setTodoPhases(phases);
-	runtime.sessionManager.appendCustomEntry(USER_TODO_EDIT_CUSTOM_TYPE, { phases });
-}
-
-export function copyPhases(phases: TodoPhase[]): TodoPhase[] {
-	return phases.map(phase => ({ ...phase, tasks: phase.tasks.slice() }));
-}
