+++ packages/coding-agent/src/slash-commands/helpers/format.ts
@@ -1,3 +1,4 @@
+/** Format a millisecond duration as a coarse-grained human label. */
 export function formatDuration(ms: number): string {
 	const seconds = Math.max(0, Math.round(ms / 1000));
 	if (seconds < 60) return `${seconds}s`;
@@ -9,6 +10,10 @@ export function formatDuration(ms: number): string {
 	return `${days}d`;
 }
 
+/**
+ * Render an ASCII progress bar with a trailing percent label.
+ * `fraction` is clamped to `[0, 1]`. `undefined` renders a dotted placeholder.
+ */
 export function renderAsciiBar(fraction: number | undefined, width = 24): string {
 	if (fraction === undefined) return `[${"·".repeat(width)}]`;
 	const clamped = Math.min(Math.max(fraction, 0), 1);
